requirements.txt
args==0.1.0
attrs==18.2.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
click-completion==0.5.2
click-didyoumean==0.0.3
click-help-colors==0.6
clint==0.5.1
colorama==0.3.9
cryptography==2.8
cursor==1.3.4
cycler==0.10.0
future==0.18.2
gradient==0.5.2
graphviz==0.13.2
halo==0.0.28
idna==2.8
Jinja2==2.11.1
kiwisolver==1.1.0
log-symbols==0.0.14
MarkupSafe==1.1.1
marshmallow==2.20.5
matplotlib==3.1.3
numpy==1.18.1
opencv-contrib-python==4.2.0.32
pathlib2==2.3.5
Pillow==7.0.0
progressbar2==3.47.0
prompt-toolkit==2.0.10
pycparser==2.19
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2019.3
PyYAML==5.3
raven==6.10.0
requests==2.22.0
requests-toolbelt==0.9.1
shellingham==1.3.1
six==1.14.0
spinners==0.0.23
tabulate==0.8.6
termcolor==1.1.0
terminaltables==3.1.0
torch==1.5.0
torchvision==0.6.0
torchviz==0.0.1
urllib3==1.25.8
wcwidth==0.1.8
